Add mirrors.kernel.org entries
There are multiple nodes serving mirrors.kernel.org, with one of them
only serving rsync, and others only serving https. Normally, they are
accessed as mirrors.edge.kernel.org with geoDNS magic routing the
request to the nearest node, but it's possible to localize down to a
specific geographical area.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>