Improve subnet checking logic
This change significantly improves performance and decreases underlying
complexity of the subnet matching system.  It offloads the matching work
mostly to the mirror status checker so we're only doing basic integer
comparisons for live non-cached requests instead of having the ipaddress
module have to do all of the logic for every non-cached request.

Some basic profiling data before/after of a non-cached request:
640542 function calls (637832 primitive calls) in 0.471 seconds
168042 function calls (165336 primitive calls) in 0.160 seconds

This is about a 3x speedup for non-cached requests that will yield
faster responses to users and resource savings for the mirror system.